Upon service reset, also clear the cached token
If a cache is set then a token is acquired it is stored in both the Cache service and the Properties service. Therefore, when the Service is reset the token must be removed from the Cache service as well as the Properties service. Otherwise, getAccessToken() would still return the cached token after a call to reset().