Add function to return id_token (for Service Now)
Some IDPs such as OpenAM can be configured to provide both an access_token and an id_token with the callback.  This update provides a function that can be called to return the id_token instead of the access_token.  The id_token contains a payload with claims related to the authenticated identity.  This was required for API integration to Service Now, who told me their service cannot be configured to accept an access_token.